Fix: Greasy feet were not wearing off. Fixes #135.
Getting fumbling from a grease trap would incorrectly set the timeout for the fumbling to 0, causing the timeout to never trigger and making the player permanently greased. Very unfortunate. I noticed I didn't use longs where I should have, probably resulting in the 0 init value.

I believe this also fixed 135 because if the value was 0 it makes sense that when applying a towel the check for fumbling might not evaluate correctly.

+++ include/extern.h
@@ -2587,7 +2587,7 @@ extern void toggle_blindness(void);
 extern boolean make_hallucinated(long, boolean, long);
 extern void make_deaf(long, boolean);
 extern void make_glib(int);
-extern void make_fumbling(int);
+extern void make_fumbling(long);
 extern void self_invis_message(void);
 extern int dodrink(void);
 extern int dopotion(struct obj *) NONNULLARG1;
